+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Directory of User and My Profile Data Improvements [#4865](https://github.com/rokwire/illinois-app/issues/4865).
+### Changed
+- GroupAllEventsPanel extended to work with upcoming and past events [#5022](https://github.com/rokwire/illinois-app/issues/5022).
 
 ## [7.0.16] - 2025-04-23
 ### Changed